update: fixed ui for backup screen
