+++ app/src/main/java/org/satochip/seedkeeper/ui/components/backup/MainBackupButton.kt
@@ -18,26 +18,21 @@ fun MainBackupButton(
                 BackupStatus.DEFAULT -> {
                     backupStatus.value = BackupStatus.FIRST_STEP
                 }
-
                 BackupStatus.FIRST_STEP -> {
                     backupStatus.value = BackupStatus.SECOND_STEP
                     showNfcDialog.value = !showNfcDialog.value
                 }
-
                 BackupStatus.SECOND_STEP -> {
                     backupStatus.value = BackupStatus.THIRD_STEP
                     showNfcDialog.value = !showNfcDialog.value
                 }
-
                 BackupStatus.THIRD_STEP -> {
                     backupStatus.value = BackupStatus.FOURTH_STEP
                 }
-
                 BackupStatus.FOURTH_STEP -> {
                     backupStatus.value = BackupStatus.FIFTH_STEP
                     showNfcDialog.value = !showNfcDialog.value
                 }
-
                 BackupStatus.FIFTH_STEP -> {
                     onClick()
                 }
@@ -47,23 +42,18 @@ fun MainBackupButton(
             BackupStatus.DEFAULT -> {
                 R.string.start
             }
-
             BackupStatus.FIRST_STEP -> {
                 R.string.next
             }
-
             BackupStatus.SECOND_STEP -> {
                 R.string.scanMySeedkeeper
             }
-
             BackupStatus.THIRD_STEP -> {
                 R.string.makeBackup
             }
-
             BackupStatus.FOURTH_STEP -> {
                 R.string.next
             }
-
             BackupStatus.FIFTH_STEP -> {
                 R.string.home
             }
